refactor: Extract confirmation popup code to its own file
